pyproject.toml
[tool.poetry]
name = "financeiro-pessoal"
version = "0.1.0"
description = ""
authors = ["Angelo Velloso"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic-settings = "^2.1.0"
sling = "^1.1.5.post4"
streamlit = "^1.31.1"
tabula-py = {extras = ["jpype"], version = "^2.9.0"}
pdfquery = "^0.4.3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.2.1"
httpx = "^0.26.0"
isort = "^5.13.2"
ignr = "^2.2"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff && isort --check . --diff'
format = 'blue .  && isort .'
api = 'cd app && uvicorn api:app --reload'
ui = 'streamlit run app/Home.py'
#pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'